# Optimise progress bars on course page