--- a/app/controllers/courses_controller.rb
+++ b/app/controllers/courses_controller.rb
@@ -91,6 +91,11 @@ class CoursesController < ApplicationController
     @title = @course.name
     @series = policy_scope(@course.series).includes(:evaluation)
     @series_loaded = params[:secret].present? ? @course.series.count : 2
+    @progress_info = {
+      users_tried_by_exercise: @course.users_tried_by_exercise,
+      users_correct_by_exercise: @course.users_correct_by_exercise,
+      users_read_by_content_page: @course.users_read_by_content_page
+    }
   end
 
   # GET /courses/new