--- a/app/models/submission.rb
+++ b/app/models/submission.rb
@@ -60,7 +60,7 @@ class Submission < ApplicationRecord
   scope :of_judge, ->(judge) { where(exercise_id: Exercise.where(judge_id: judge.id)) }
   scope :from_students, ->(course) { where(user: course.enrolled_members) }
 
-  scope :judged, -> { where.not(status: %i[running queued]) }
+  scope :judged, -> { where(status: ['unknown', 'correct', 'wrong', 'time limit exceeded', 'runtime error', 'compilation error', 'memory limit exceeded', 'internal error', 'output limit exceeded']) }
   scope :by_exercise_name, ->(name) { where(exercise: Exercise.by_name(name)) }
   scope :by_status, ->(status) { where(status: status.in?(statuses) ? status : -1) }
   scope :by_username, ->(name) { where(user: User.by_filter(name)) }