--- a/app/views/courses/show.html.erb
+++ b/app/views/courses/show.html.erb
@@ -131,7 +131,7 @@
     <% end %>
     <div id="series-listing">
       <% @series.each do |series| %>
-        <%= render(partial: 'series/series', locals: {series: series, progress_info: @progress_info}) %>
+        <%= render(partial: 'series/series', locals: {series: series }) %>
       <% end %>
     </div>
   </div>