… app/views/series/_series.html.erb
… app/views/series/_series_status.html.erb
… config/environments/development.rb
… db/migrate/20221116145957_add_course_progress_index_to_submissions.rb
… db/schema.rb
… test/factories/courses.rb
… test/factories/series.rb
… test/fixtures/courses.yml
… test/models/course_test.rb
… test/models/series_test.rb