This pull request optimizes the queries to load progress bars on the course page. This removes the need for partially loading the page as the slow elements are removed. All code for the partial load is thus also removed.

To optimize the queries I introduced an index. I also had to update the `judged` check for submission from `NOT IN` to  `IN` (With the list of statuses inverted). Without this change mysql refused to use the provided index. 

Replaces #4167 

Below you can find the results of a speed comparison. For fair comparison I cleared my sql cache before each request. I tested on https://dodona.ugent.be/nl/courses/359/ which is a rather large course.
- Old: 2421ms/1798ms/1887ms
- Load course with all series: 14446ms/13961ms
- Load course with only progress bars hidden: 2445ms/2345ms/2298ms/2513ms
-  #4167: 1875ms/1900ms/1991ms
- New: 1027ms/1049ms/918ms/876ms/1060ms 

Closes https://github.com/dodona-edu/dodona/issues/3962 .
Closes: #4167 

PS: this pr also contained some changes to the generated database file.
This is probably caused by me having a backup of production on my machine.
I would keep these changes in as they reflect the state of the production database.